+++ src/ngraph/frontend/onnx_import/core/attribute.cpp
