+++ src/ngraph/frontend/onnx_import/op/batch_norm.hpp
@@ -18,8 +18,8 @@
 
 #include <memory>
 
+#include "core/node.hpp"
 #include "exceptions.hpp"
-#include "wrappers/node.hpp"
 
 #include "ngraph/node_vector.hpp"
 #include "ngraph/op/batch_norm.hpp"
